+++ src/ru/smg/kanban/tasks/Epic.java
@@ -7,31 +7,32 @@ public class Epic extends Task {
     public ArrayList<Subtask> getSubtasks() {
         return subtasks;
     }
+
     private final ArrayList<Subtask> subtasks;
 
-    public Epic(long id, String name, String description, ArrayList<Subtask> subtasks) {
+    public Epic(int id, String name, String description, ArrayList<Subtask> subtasks) {
         super(id, name, description, Status.NEW);
         this.subtasks = subtasks;
     }
 
     @Override
-    public Status getStatus(){
-        if (subtasks.size() == 0){
+    public Status getStatus() {
+        if (subtasks.size() == 0) {
             return Status.NEW;
         }
         boolean isDone = true;
         boolean isNew = true;
-        for (Subtask subtask: subtasks) {
-            if (subtask.getStatus() != Status.NEW){
+        for (Subtask subtask : subtasks) {
+            if (subtask.getStatus() != Status.NEW) {
                 isNew = false;
             }
-            if (subtask.getStatus() != Status.DONE){
+            if (subtask.getStatus() != Status.DONE) {
                 isDone = false;
             }
         }
-        if (isNew){
+        if (isNew) {
             return Status.NEW;
-        } else if (isDone){
+        } else if (isDone) {
             return Status.DONE;
         }
         return Status.IN_PROGRESS;
