+++ src/ru/smg/kanban/tasks/Subtask.java
@@ -5,9 +5,10 @@ public class Subtask extends Task {
     public Epic getHolder() {
         return holder;
     }
+
     private final Epic holder;
 
-    public Subtask(long id, String name, String description, Status status, Epic holder) {
+    public Subtask(int id, String name, String description, Status status, Epic holder) {
         super(id, name, description, status);
         this.holder = holder;
     }
