+++ src/ru/smg/kanban/tasks/Task.java
@@ -2,9 +2,10 @@ package ru.smg.kanban.tasks;
 
 public class Task {
 
-    public long getId() {
+    public int getId() {
         return id;
     }
+
     public String getName() {
         return name;
     }
@@ -17,12 +18,12 @@ public class Task {
         return status;
     }
 
-    private final long id;
+    private final int id;
     private final String name;
     private final String description;
     private final Status status;
 
-    public Task(long id, String name, String description, Status status ) {
+    public Task(int id, String name, String description, Status status) {
         this.id = id;
         this.name = name;
         this.description = description;
